+++ mayaTools/cgm/core/cgmPy/os_Utils.py
@@ -15,6 +15,7 @@ Key:
 import re
 import os
 import stat
+import pprint
 
 #>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 import logging
@@ -36,6 +37,11 @@ from cgm.core import cgm_General as cgmGEN
 #reload(zooPath)
 #>>> Utilities
 #===================================================================
+
+log_start = cgmGEN.logString_start
+log_sub = cgmGEN.logString_sub
+log_msg = cgmGEN.logString_msg
+
 def get_lsFromPath(str_path = None, 
                    matchArg = None, 
                    calledFrom = None,
@@ -411,6 +417,223 @@ def verify_dir_fromDict(root = None, d = {}, case = None):
                 log.warning("created dir: {0}".format(mSub))                
             
         
+def find_emptyDirs(path = None, 
+                   addFile = False,
+                   delEmpty = False,
+                   level = None,):
+    """
+    Function for walking below a given directory looking for empty directories
+    
+    :parameters
+        path(str)
+        level(int) - Depth to search. None means everything
+        mode(int)
+            0 - normal
+            1 - pycs only
+        self(instance): cgmMarkingMenu
+        cleanPyc: Delete pycs after check
+    :returns
+        _d_files,_l_ordered,_l_pycd
+        _d_files - dict of import key to file
+        _l_ordered - ordered list of module keys as found
+        _l_pycd - list of modules that were _pycd
+    
+    """
+    _str_func = 'find_emptyDirs'
+    _b_debug = log.isEnabledFor(logging.DEBUG)
 
+    _path = PATH.Path(path)
+    
+    _l_subs = []
+    _d_files = {}
+    _d_names = {}
+    _l_empty = []
+    _l_new = []
+    _l_removed = []
+    
+    _base = _path.split()[-1]
+    
+    log.debug("|{0}| >> Checking base: {1} | path: {2}".format(_str_func,_base,path))
+    
+    _i = 0
+    for root, dirs, files in os.walk(path, True, None):
+        
+        # Parse all the files of given path and reload python modules
+        _mRoot = PATH.Path(root)
+        _split = _mRoot.split()
+        _subRoot = _split[-1]
+        _splitUp = _split[_split.index(_base):]
+        
+        log.debug("|{0}| >> On subroot: {1} | path: {2}".format(_str_func,_subRoot,root))   
+        log.debug("|{0}| >> On split: {1}".format(_str_func,_splitUp))
+        
+        log.debug( dirs )
+        log.debug(files )
+        
+        if not dirs and not files:
+            _l_empty.append(root)
+            
+            if delEmpty:
+                os.rmdir(root)
+                _l_removed.append(root)
+                continue
+            
+            if addFile:
+                _new = os.path.join(root,"empty.txt")
+                #open(_new,'x')
+                os.close(os.open(_new, os.O_CREAT|os.O_EXCL))#...from stack https://stackoverflow.com/questions/43081859/how-to-create-empty-file-in-python
+                _l_new.append(_new)
+
+
+        if level is not None and _i >= level:break 
+        _i +=1
+        
+    if _l_empty and not addFile:
+        log.info(log_sub(_str_func,"Empty..."))
+        pprint.pprint(_l_empty)
+        log.info(cgmGEN._str_subLine)
+    if _l_new:
+        log.info(log_sub(_str_func,"New..."))        
+        pprint.pprint(_l_new)
+        log.info(cgmGEN._str_subLine)
+        
+    if _l_removed:
+        log.info(log_sub(_str_func,"Removed..."))        
+        pprint.pprint(_l_removed)
+        log.info(cgmGEN._str_subLine)        
+        
+    return
  
+def find_tmpFiles(path = None, level = None, cleanFiles = False,
+                  endMatch = ['_batch.py','_MRSbatch.py'],
+                  l_mask = ['max',
+                            'mab',
+                            'markdown',
+                            'mapping']):
+    """
+    Function for walking below a given directory looking for modules to reload. It finds modules that have pyc's as
+    well for help in reloading. There is a cleaner on it as well to clear all pycs found.
+    
+    :parameters
+        path(str)
+        level(int) - Depth to search. None means everything
+        mode(int)
+            0 - normal
+            1 - pycs only
+        self(instance): cgmMarkingMenu
+        cleanPyc: Delete pycs after check
+    :returns
+        _d_files,_l_ordered,_l_pycd
+        _d_files - dict of import key to file
+        _l_ordered - ordered list of module keys as found
+        _l_pycd - list of modules that were _pycd
+    
+    """
+    _str_func = 'find_tmpFiles'
+    _b_debug = log.isEnabledFor(logging.DEBUG)
 
+    _path = PATH.Path(path)
+    
+    _l_subs = []
+    _d_files = {}
+    _d_names = {}
+    
+    _l_duplicates = []
+    _l_errors = []
+    _base = _path.split()[-1]
+    _l_ordered_list = []
+    _l_weirdFiles = []
+    _d_weirdFiles = {}
+    
+    log.debug("|{0}| >> Checking base: {1} | path: {2}".format(_str_func,_base,path))
+    
+    _i = 0
+    for root, dirs, files in os.walk(path, True, None):
+        
+        # Parse all the files of given path and reload python modules
+        _mRoot = PATH.Path(root)
+        _split = _mRoot.split()
+        _subRoot = _split[-1]
+        _splitUp = _split[_split.index(_base):]
+        
+        log.debug("|{0}| >> On subroot: {1} | path: {2}".format(_str_func,_subRoot,root))   
+        log.debug("|{0}| >> On split: {1}".format(_str_func,_splitUp))   
+        
+        _mod = False
+        _l_sub = []
+        for f in files:
+            key = False
+            _pycd = False
+            _long = os.path.join(root,f)
+            
+            if '.' not in f:
+                continue
+            
+            _dot_split = f.split('.')
+            _extension = _dot_split[-1]
+            _pre = _dot_split[0]
+            
+            if _extension in l_mask:
+                continue
+            
+            if len(_extension) > 3:
+                if _extension.startswith('ma') or _extension.startswith('mb'):
+                    _l_weirdFiles.append(f)
+                    _d_weirdFiles[f] = os.path.join(root,f)
+                    continue
+            
+            for s in endMatch:
+                if f.endswith(s):
+                    _l_weirdFiles.append(f)
+                    _d_weirdFiles[f] = os.path.join(root,f)
+                    continue                
+            
+
+        if level is not None and _i >= level:break 
+        _i +=1
+        
+    if cleanFiles:
+        for f,_path in _d_weirdFiles.iteritems():
+            try:
+                log.warning("Remove: {0}".format(_path))
+                os.remove( _path )
+            except WindowsError, e:
+                try:
+                    log.info("|{0}| >> Initial delete fail. attempting chmod... ".format(_str_func))                          
+                    os.chmod( _path, stat.S_IWRITE )
+                    os.remove( _path )                          
+                except Exception,e:
+                    for arg in e.args:
+                        log.error(arg)   
+                    raise RuntimeError,"Stop"
+    else:
+        if _d_weirdFiles:
+            pprint.pprint(_d_weirdFiles)
+            log.warning( cgmGEN.logString_msg(_str_func,"Found {0} files".format(len(_d_weirdFiles.keys()))) )
+        else:
+            log.warning( cgmGEN.logString_msg(_str_func,"No files found.") )
+    return
+    """
+    if cleanPyc:
+        _l_failed = []
+        log.debug("|{0}| >> Found {1} pyc files under: {2}".format(_str_func,len(_l_pyc),path))                        
+        for _file in _l_pyc:
+        #for k in _l_ordered_list:
+            #if k in _l_pycd:
+            log.debug("|{0}| >> Attempting to clean pyc for: {1} ".format(_str_func,_file))  
+            if not _file.endswith('.pyc'):
+                raise ValueError,"Should NOT be here"
+            try:
+                os.remove( _file )
+            except WindowsError, e:
+                try:
+                    log.info("|{0}| >> Initial delete fail. attempting chmod... ".format(_str_func))                          
+                    os.chmod( _file, stat.S_IWRITE )
+                    os.remove( _file )                          
+                except Exception,e:
+                    for arg in e.args:
+                        log.error(arg)   
+                    raise RuntimeError,"Stop"
+        _l_pyc = []
+        """
+    #return _d_files, _l_ordered_list, _l_pycd
