+++ mayaTools/cgm/core/mrs/lib/scene_utils.py
@@ -26,14 +26,15 @@ from Red9.core import Red9_AnimationUtils as r9Anim
 import logging
 logging.basicConfig()
 log = logging.getLogger(__name__)
-log.setLevel(logging.INFO)
+log.setLevel(logging.DEBUG)
 #========================================================================
-__version__ = '1.05312019'
+__version__ = cgmGEN.__RELEASE
 
 import maya.cmds as mc
 from cgm.core import cgm_General as cgmGEN
 
 from cgm.core.cgmPy import path_Utils as PATH
+import cgm.core.cgmPy.os_Utils as CGMOS
 
 import cgm.core.classes.GuiFactory as cgmUI
 mUI = cgmUI.mUI
@@ -83,144 +84,13 @@ d_annotations = {
 'saveHere':'Open maya dialog to open file here',
 
 }
-
+log_start = cgmGEN.logString_start
+log_sub = cgmGEN.logString_sub
+log_msg = cgmGEN.logString_msg
 #=============================================================================================================
 #>> Queries
 #=============================================================================================================
-def find_tmpFiles(path = None, level = None, cleanFiles = False,
-                  endMatch = ['_batch.py','_MRSbatch.py'],
-                  l_mask = ['max',
-                            'mab',
-                            'markdown',
-                            'mapping']):
-    """
-    Function for walking below a given directory looking for modules to reload. It finds modules that have pyc's as
-    well for help in reloading. There is a cleaner on it as well to clear all pycs found.
-    
-    :parameters
-        path(str)
-        level(int) - Depth to search. None means everything
-        mode(int)
-            0 - normal
-            1 - pycs only
-        self(instance): cgmMarkingMenu
-        cleanPyc: Delete pycs after check
-    :returns
-        _d_files,_l_ordered,_l_pycd
-        _d_files - dict of import key to file
-        _l_ordered - ordered list of module keys as found
-        _l_pycd - list of modules that were _pycd
-    
-    """
-    _str_func = 'find_tmpFiles'
-    _b_debug = log.isEnabledFor(logging.DEBUG)
-
-    _path = PATH.Path(path)
-    
-    _l_subs = []
-    _d_files = {}
-    _d_names = {}
-    
-    _l_duplicates = []
-    _l_errors = []
-    _base = _path.split()[-1]
-    _l_ordered_list = []
-    _l_weirdFiles = []
-    _d_weirdFiles = {}
-    
-    log.debug("|{0}| >> Checking base: {1} | path: {2}".format(_str_func,_base,path))
-    
-    _i = 0
-    for root, dirs, files in os.walk(path, True, None):
-        
-        # Parse all the files of given path and reload python modules
-        _mRoot = PATH.Path(root)
-        _split = _mRoot.split()
-        _subRoot = _split[-1]
-        _splitUp = _split[_split.index(_base):]
-        
-        log.debug("|{0}| >> On subroot: {1} | path: {2}".format(_str_func,_subRoot,root))   
-        log.debug("|{0}| >> On split: {1}".format(_str_func,_splitUp))   
-        
-        _mod = False
-        _l_sub = []
-        for f in files:
-            key = False
-            _pycd = False
-            _long = os.path.join(root,f)
-            
-            if '.' not in f:
-                continue
-            
-            _dot_split = f.split('.')
-            _extension = _dot_split[-1]
-            _pre = _dot_split[0]
-            
-            if _extension in l_mask:
-                continue
-            
-            if len(_extension) > 3:
-                if _extension.startswith('ma') or _extension.startswith('mb'):
-                    _l_weirdFiles.append(f)
-                    _d_weirdFiles[f] = os.path.join(root,f)
-                    continue
-            
-            for s in endMatch:
-                if f.endswith(s):
-                    _l_weirdFiles.append(f)
-                    _d_weirdFiles[f] = os.path.join(root,f)
-                    continue                
-            
-                    
-
-        if level is not None and _i >= level:break 
-        _i +=1
-        
-    if cleanFiles:
-        for f,_path in _d_weirdFiles.iteritems():
-            try:
-                log.warning("Remove: {0}".format(_path))
-                os.remove( _path )
-            except WindowsError, e:
-                try:
-                    log.info("|{0}| >> Initial delete fail. attempting chmod... ".format(_str_func))                          
-                    os.chmod( _path, stat.S_IWRITE )
-                    os.remove( _path )                          
-                except Exception,e:
-                    for arg in e.args:
-                        log.error(arg)   
-                    raise RuntimeError,"Stop"
-    else:
-        if _d_weirdFiles:
-            pprint.pprint(_d_weirdFiles)
-            log.warning( cgmGEN.logString_msg(_str_func,"Found {0} files".format(len(_d_weirdFiles.keys()))) )
-        else:
-            log.warning( cgmGEN.logString_msg(_str_func,"No files found.") )
-    return
-    """
-    if cleanPyc:
-        _l_failed = []
-        log.debug("|{0}| >> Found {1} pyc files under: {2}".format(_str_func,len(_l_pyc),path))                        
-        for _file in _l_pyc:
-        #for k in _l_ordered_list:
-            #if k in _l_pycd:
-            log.debug("|{0}| >> Attempting to clean pyc for: {1} ".format(_str_func,_file))  
-            if not _file.endswith('.pyc'):
-                raise ValueError,"Should NOT be here"
-            try:
-                os.remove( _file )
-            except WindowsError, e:
-                try:
-                    log.info("|{0}| >> Initial delete fail. attempting chmod... ".format(_str_func))                          
-                    os.chmod( _file, stat.S_IWRITE )
-                    os.remove( _file )                          
-                except Exception,e:
-                    for arg in e.args:
-                        log.error(arg)   
-                    raise RuntimeError,"Stop"
-        _l_pyc = []
-        """
-    #return _d_files, _l_ordered_list, _l_pycd
+find_tmpFiles = CGMOS.find_tmpFiles
     
 def buildMenu_utils(self, mMenu):
 
@@ -266,6 +136,23 @@ def buildMenu_utils(self, mMenu):
     
     
     mUI.MelMenuItemDiv( mMenu )
+    
+    #Empty
+    _empty = mUI.MelMenuItem(mMenu,l='Empty Dirs',subMenu=True)
+    
+    mUI.MelMenuItem(_empty,
+                  label='Add empty.txt',ut='cgmUITemplate',
+                  c=lambda *a: CGMOS.find_emptyDirs( self.directory,True,False),
+                  ann='Add empty txt files')
+    
+    mUI.MelMenuItem(_empty,
+                  label='Remove',ut='cgmUITemplate',
+                   c=lambda *a: CGMOS.find_emptyDirs( self.directory,False,True),
+                   ann='Delete empty dirs')    
+    
+    
+    
+    
     
     #DropBox...
     _fileTrash = mUI.MelMenuItem(mMenu,l='File Trash',subMenu=True)
@@ -375,4 +262,6 @@ def fncMayaSett_query(self):
                 else:
                     log.warning("No function found for {0} | {1}".format(k,_name))
             except Exception,err:
-                log.error("Failure {0} | {1} | {2}".format(k,_name,err))        
+                log.error("Failure {0} | {1} | {2}".format(k,_name,err))
+                
+
